resume-platform/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.8.2
python-dotenv==1.0.0
reportlab==4.0.7
email-validator==2.3.0
